+++ stf-api/extensions/users-permissions/config/jwt.js
@@ -1,3 +1,0 @@
-module.exports = {
-  jwtSecret: process.env.JWT_SECRET || '2ecb3e1f-2d54-4255-b551-f55776be6ab3'
-};