+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.1] - 26/8/21
+### Changed
+- order of blog posts
+
+### Added
+- Link to Ornate frame component
 ## [1.1.0] - 18/8/21
 ### Changes
 - Full website live!